Increase dog's armor to make it harder to kill
Armor does not mean real armor. It's just the percentage of damage that the mob can get.

Say a mob was punched with 50 damage, and its armor was 30. Then the mob would only take 15 damage.

This makes it more reasonable; while the dog has the same amount of health, it is more resistant to enemies.